// src/scripts/csv/types/Supplier.ts
export interface ISupplierCSV {
  provider_id: string;
  acceptsassignement: string;
  latestclaimsubmitteddate: string;
  supplieslist: string;
  providertypelist: string;
  specialitieslist: string;
  telephonenumber: string;
  practicezip9code: string;
  practicestate: string;
  practicecity: string;
  practiceaddress2: string;
  practiceaddress1: string;
  latitude: number;
  longitude: number;
  practicename: string;
  businessname: string;
  ptan: string;
  participationbegindate: string;
  is_contracted_for_cba: string;
}